Create DAY 146
3423. Maximum Difference Between Adjacent Elements in a Circular Array
🌀 Problem in Simple Words:
You’re given a list of numbers that are arranged in a circle. You have to find the biggest difference between any two numbers that are next to each other in that circle.

👉 Remember: Since it’s a circle, the last number is also next to the first one.

🧠 What to Do:
Go through each number in the list.

For every number, look at the next one. If you're at the last number, look at the first one (because it wraps around like a circle).

Find the absolute difference between the two numbers (we ignore which one is bigger).

Keep track of the biggest difference you find.

🔁 Example:
If the array is [10, 3, 6, 1], then the pairs to compare are:

|10 - 3| = 7

|3 - 6| = 3

|6 - 1| = 5

|1 - 10| = 9 ← because it's circular!

So, the answer is 9.

⏱️ Time it Takes:
Just one round through the list → O(n) time.